receive partials
